+++ service/configservice/api.go
@@ -8609,11 +8609,11 @@ func (c *ConfigService) PutConfigRuleRequest(input *PutConfigRuleInput) (req *re
 // Adds or updates an Config rule for evaluating whether your Amazon Web Services
 // resources comply with your desired configurations.
 //
-// You can use this action for custom Config rules and Config managed rules.
-// A custom Config rule is a rule that you develop and maintain. An Config managed
+// You can use this action for Config custom rules and Config managed rules.
+// A Config custom rule is a rule that you develop and maintain. An Config managed
 // rule is a customizable, predefined rule that Config provides.
 //
-// If you are adding a new custom Config rule, you must first create the Lambda
+// If you are adding a new Config custom rule, you must first create the Lambda
 // function that the rule invokes to evaluate your resources. When you use the
 // PutConfigRule action to add the rule to Config, you must specify the Amazon
 // Resource Name (ARN) that Lambda assigns to the function. Specify the ARN
@@ -9465,8 +9465,8 @@ func (c *ConfigService) PutOrganizationConfigRuleRequest(input *PutOrganizationC
 // by calling Amazon Web Services Organization register-delegated-administrator
 // for config-multiaccountsetup.amazonaws.com.
 //
-// You can use this action to create both custom Config rules and Config managed
-// rules. If you are adding a new custom Config rule, you must first create
+// You can use this action to create both Config custom rules and Config managed
+// rules. If you are adding a new Config custom rule, you must first create
 // Lambda function in the master account or a delegated administrator that the
 // rule invokes to evaluate your resources. You also need to create an IAM role
 // in the managed-account that can be assumed by the Lambda function. When you
@@ -31987,6 +31987,12 @@ const (
 
 	// ResourceTypeAwsGuardDutyDetector is a ResourceType enum value
 	ResourceTypeAwsGuardDutyDetector = "AWS::GuardDuty::Detector"
+
+	// ResourceTypeAwsEmrSecurityConfiguration is a ResourceType enum value
+	ResourceTypeAwsEmrSecurityConfiguration = "AWS::EMR::SecurityConfiguration"
+
+	// ResourceTypeAwsSageMakerCodeRepository is a ResourceType enum value
+	ResourceTypeAwsSageMakerCodeRepository = "AWS::SageMaker::CodeRepository"
 )
 
 // ResourceType_Values returns all elements of the ResourceType enum
@@ -32110,6 +32116,8 @@ func ResourceType_Values() []string {
 		ResourceTypeAwsEc2LaunchTemplate,
 		ResourceTypeAwsEcrPublicRepository,
 		ResourceTypeAwsGuardDutyDetector,
+		ResourceTypeAwsEmrSecurityConfiguration,
+		ResourceTypeAwsSageMakerCodeRepository,
 	}
 }